backend/requirements.txt
google-generativeai
requests
Flask